+++ public/collection/algorithms/dfs.js
@@ -1,5 +1,5 @@
 /**
- * Depth first search: works for both undirected and directed graphs
+ * Depth first search: works for both undirected and directed graphs.
  */
 
 // time at which each node is discovered
@@ -20,10 +20,15 @@ step(() => {
     hideAllEdgeWeights();
 })
 
-let start = promptNode("Enter start node:");
-visit(start);
+// let user choose another starting node if there are remaining unreachable nodes
+let unvisited = new Set(getNodes());
+while ( unvisited.size > 0 ) {
+    let start = promptNodeFrom("Enter start node:", unvisited);
+    visit(start);
+}
 
 function visit(node) {
+    unvisited.delete(node);
     display(`visit ${node}`);
     step(() => {
         discoveryTimes[node] = time++;
@@ -32,9 +37,9 @@ function visit(node) {
         label(node, discoveryTimes[node]);
     });
 
-	for ( let edge of outgoing(node) ) {
+     for ( let edge of outgoing(node) ) {
         if ( hasColor(edge) ) continue; // seen this edge from the other end (undirected)
-	    let nextNode = other(node, edge);
+           let nextNode = other(node, edge);
             if ( hasLabel(edge) ) {
                 continue;
             }
@@ -44,7 +49,6 @@ function visit(node) {
                 highlight(edge);
                 color(edge, "blue");
                 highlight(nextNode);
-                visit(nextNode);
             } else if ( finishTimes[nextNode] == null ) { // ancestor
                 label(edge, "B");
                 color(edge, "red");
@@ -56,6 +60,10 @@ function visit(node) {
                 color(edge, "orange");
             }
         });
+        // keep the recursive call outside of the step
+        if ( ! marked(nextNode) ) { // not yet visited
+            visit(nextNode);
+        }
     }
 
     finishTimes[node] = time++;