+++ public/collection/algorithms/quicksort.js
@@ -2,7 +2,6 @@
  * An animation of Quicksort based on a project
  * submitted by Hayden Fuss, Solomon Yeh, and Jordan Connor
  * in CSC 316, Spring 2015.
- * Works for sorting graphs only.
  */
 
 // Algorithm
@@ -16,11 +15,6 @@
 //      greater = quicksort(greater)
 //      return append less, equal, greater
 
-Algorithm.configure({
-    controlNodePosition: true
-})
-
-// @todo should not need these
 const HORIZONTAL_GAP = 1;
 const VERTICAL_GAP = 1;
 
@@ -44,7 +38,7 @@ function lineUpNodes(nodeList) {
         for ( const node of nodeIds ) {
     //      to avoid multiple steps, we need to do this "manually"
     //        placeNode(node, level, index, "white", "circle")
-            setPosition(node, {x: HORIZONTAL_GAP * index, y: 2 * level * VERTICAL_GAP})
+            setPosition(node, HORIZONTAL_GAP * index, 2 * level * VERTICAL_GAP)
             color(node, "white")
             setShape(node, "circle")
             index++;
@@ -55,7 +49,7 @@ function lineUpNodes(nodeList) {
 function placeNode(node, level, index, desiredColor, shape) {
     step(()=> {
 //        display(`-> placeNode: ${node}, ${level}, ${index}`)
-        setPosition(node, {x: HORIZONTAL_GAP * index, y: 2 * level * VERTICAL_GAP})
+        setPosition(node, HORIZONTAL_GAP * index, 2 * level * VERTICAL_GAP)
         color(node, desiredColor)
         setShape(node, shape)
     })
@@ -64,7 +58,7 @@ function placeNode(node, level, index, desiredColor, shape) {
 function placePivot(pivot, level, index) {
     step(() => {
 //        display(`-> placePivot: ${pivot}, ${level}, ${index}`)
-        setPosition(pivot, {x: HORIZONTAL_GAP * index, y: (2 * level - 1) * VERTICAL_GAP})
+        setPosition(pivot, HORIZONTAL_GAP * index, (2 * level - 1) * VERTICAL_GAP)
         color(pivot, "black")
         setShape(pivot, "triangle")
     })
@@ -73,7 +67,7 @@ function placePivot(pivot, level, index) {
 function moveToSorted(node) {
     step(() => {
 //        display(`-> moveToSorted: ${node}`)
-        incrementPosition(node, {x: 0, y: -2 * VERTICAL_GAP})
+        incrementPosition(node, 0, -2 * VERTICAL_GAP)
         color(node, "yellow")
         setShape(node, "star")
     })
@@ -141,7 +135,6 @@ function quicksort(list, left, right, depth) {
 let nodeIds = getNodes();
 
 step(() => {
-    showAllNodeWeights();
     clearNodeColors();
     clearNodeShapes();
 })